README: add ~/.local/bin to path if needed
stock ubuntu .bashrc will add ~/.local/bin (and ~/bin) to the PATH
but only if they exist at login time.

On a new VM ./local/bin is probably not in the PATH so test and add it if
need be. (Don't source the whole setup script as that makes things messy.)

Signed-off-by: Bill Mills <bill.mills@linaro.org>