remote-build: add and disable other layer updates
The idea was to push local changes in layers to the remote host before
build.

This currently does not work so is disabled.

The current git push strategy requires things to be committed at least
locally.   kas will not overwrite local non-committed changes but will
undo any committed changes.  So the two are incompatible.

Signed-off-by: Bill Mills <bill.mills@linaro.org>