generic-armv7a: don't add firmware to rootfs by default.
The default firmware package for this config is huge.
We don't really need most of it.
Including it in the rootfs makes the cpio file too big to fit in
STMP157x-DK2 memory.

For now leave it out and will will add back anything needed on deploy
to test.

One could argue we don't need most of the kernel modules but
they are OK for now.

Signed-off-by: Bill Mills <bill.mills@linaro.org>