Add linux-openamp staging kernel
linux-openam-staging is a kernel repo that holds all kernel tags from:
* Linus
* Stable
* RT
* RT stable

In addition it holds hand curated branches that have include upstream
pending work that the OpenAMP choses to include for its CI purposes.

As the intention of this kernel provide is to easily try lots of different
kernel versions w/o a ton of editing, we will skip some of the pedantic
tests and trust the user input.

This also adds zynqmp-extras.cfg to this bsp layer

This commit sets up the infrastucture to add any number of extra configs
for arm64 and arm32 (v7a).  However at this time zynqmp for arm64 is the
only extra config needed.

The zynqmp extra config is needed because the arm64 defconfig does not enable
enough to do basic things on a zynqmp based board.

Arm64 defconfig does select the ARCH_ZYNQMP but that does not cascade down
to enable basic things like DMA. INTC, GPIO, SPI, Ethernet, etc.

The empty vfat.scc file eliminates a warning about the vfat feature.
The yocto kernel tools alwways add this feature and gives a warning if
not found.

vfat support is already in upstream arm64 defconfig
and arm(32) multi_v7_defconfig so nothing is needed in this file.

Signed-off-by: Bill Mills <bill.mills@linaro.org>