ci/openamp: select AUTOREV -dev versions
By default in CI, we want to test tip.

meta-openamp has introduced -dev versions of libmetal
and open-amp for this reason and to avoid AUTOREV in
the default case.  We need to opt into these.

Also in meta-openamp, the rpmgs-utils are being moved
to the openamp-system-reference repo instead of being
checked-in to the meta data.  This work is currently
a PR so we pull meta-openamp from the fork branch
until it is accepted.

* use the -dev versions of libmetal and open-amp
* switch meta-openamp to wam/wip-rpmsg-utils2
* show how to override for a specific recipe
* set all rpmsg samples to AUTOREV

Signed-off-by: Bill Mills <bill.mills@linaro.org>