Add openamp-initramfs-minimal image and switch to it.
Create an image that knows it is intended to be an initramfs image.
Therefore we can avoid pulling in a kernel image and boot etc.

We use PACKAGE_INSTALL to have more control of what gets installed.
In local.conf move the extra packages to PACKAGE_INSTALL_EXTRA.
Include the PACKAGE_INSTALL_EXTRA in IMAGE_INSTALL in case the user
builds a real image like openamp-image-minimal.

This version uses PACKAGE_EXCLUDE to
* not include any firmware by default.
* not include a kernel image
    or other boot assets that will already be loaded
* not include any kernel modules by default.

The user is expected to add any needed modules from the modules.tgz
at deployment time.

Signed-off-by: Bill Mills <bill.mills@linaro.org>