xilinx-vendor: trim the deploy dir to eliminate large redundant files
The default Xilinx build builds each filesystem image in a lot of formats.
Keep the compressed version of each *.wic.xz, *.cpio.gz, and *.tar.gz.
Remove the rest.

On disk many of these removed files are sparse files and don't take up
as much space as you would think.  However it is hard to preserve this
sparseness as files get transported and archived etc.
The compressed versions are smaller and easier to deal with.

Signed-off-by: Bill Mills <bill.mills@linaro.org>